MM-26797: Fix context cancellation on goroutines
We passed several contexts to goroutines. Therefore when the parent
function returned, the context got cancelled and therefore the running
goroutines would throw an error. To fix this:

- We remove passing the context in checkIfNeedCherryPick. This is not
ideal and every call should take the parent context. The problem is that
a request handler is semi-synchronous. It does some jobs synchronously and
spins of time taking jobs in goroutines without any control of how to shut
them down.

What we need is to make the entire event handler async, and create a context
inside that and run everything synchronously inside that. Until we have that,
we need to make do with this.

- We make createCLAPendingStatus synchronous. This also fixes a long-standing
bug where the CLA check would randomly go to a pending state.

- We move sendGitHubComment to a closure and create a new context. Again, not
ideal but have to bear with this.