Move some parts to server package
Moved transport and provider interface to the server package.

Provider interface has to be in the server package because the
interfaces are part of the consumer part instead of the implementor.

Transport has to be moved as well because if now we have a dependency
with the server in the metrics package/system and it makes sense to have
it where it's going to be used