MM-27189: Fix comment with empty body
The original code didn't comment when the error contained
"job scheduled on GitHub side; try again later"

```
if !strings.Contains("job scheduled on GitHub side; try again later", err.Error()) {
			msg := fmt.Sprintf("Error trying to update the PR.\nPlease do it manually.\nError: %s", err.Error())
			s.sendGitHubComment(ctx, pr.RepoOwner, pr.RepoName, pr.Number, msg)
		}
```

But we did not reset the err variable, which attempted to comment with an empty message.
We fix this by setting err to nil.