MM-27435: Split into a separate job server
This PR creates a separate service for the job server
and removes the functionality from the mattermod server.

A separate docker image with its own Dockerfile is provided.

The same config file needs to be used for the new service.
It is preferred that a separate GitHub token is used to separate
the services better in case one token is deleted or expired, it should
not affect the other service.