use PR from event when checking if PR is mergeable (#214)
Summary:
The production instance of mattermod seems to think that some merged and closed PRs have been closed but not merged. The closing of the PR triggers the event in mattermod, which then queries Github to check if the PR has been merged. The response from Github appears to be that the PR was not merged, although manually checking the API response from Github a little while later shows that the event has been merged.

This leads me to suspect a caching issue. In this PR mattermod uses the merge state information from the event itself instead of querying Github again. According to the Github docs [1][2], the event contains all the information we need so we should not need to query Github again.

[1] https://developer.github.com/webhooks/event-payloads/#pull_request
[2] https://developer.github.com/v3/pulls/#get-a-pull-request

Ticket Link:
N/A, but I've made a follow-up issue to fix this issue across the code base: https://mattermost.atlassian.net/browse/MM-27487