Tweak CleanOutdatedPRs (#230)
After incorporating pagination, now we have a lot of PRs in the DB.
Logs show that at the start of a run, there are 339 PRs, and by the time
the 10 minute timeout is over, there are around 200 PRs which get cancelled.

This is primarily due to the fact that there is a 5 second sleep in between.
Which allows only 12 PRs in a minute, and therefore 120 PRs in 10 minutes.
Clearly that's not enough to process 339 PRs.

To fix this, we reduce the sleep time to 200 ms to allow for 5 requests/sec,
which is well-under the burst limit of 10/sec. This will allow for 300 PRs per minute
which should now be finishable under 10 minutes.