MM-26794: Implement a DB migration plan
This PR removes all the gorp related code which would perform DB changes
automatically from models, and implements a raw SQL file based approach to migation.

We use the golang-migrate library to implement migrations. And as the first migration,
we create the base schema of the DB itself.

More info is added in the README.

A CLI tool is also added to help with rolling back schema versions.

After this, we can slowly remove the gorp code from the codebase. Now gorp does not serve
any other purpose than just an ORM.

https://mattermost.atlassian.net/browse/MM-26794