MM-27487: store PR milestone info in the DB
We store the milestone info in the DB and then use that to perform
the cherry-picking action. This saves a call to GitHub.

To achieve this, we need to differentiate between an issue and a PR
in the issue event handler because GitHub treats all PRs to be issues.

If a PR milestone is changed, we just update the DB with the latest PR info.

https://mattermost.atlassian.net/browse/MM-27487