Bump page number while continuing in case of error (#235)
* Bump page number while continuing in case of error

If a rate limit error occurs, we would just blindly retry that request
without incrementing the page number. This would keep retrying the same page
until GH would eventually return a proper response or return NextPage = 0.

A better way is to move on to the next page so that even if an error occurs,
we eventually converge to getting out of the loop.

While here, we also add some sleeps to better manage rate limit errors.
Ideally, we would just tune the burst rate, but for now we are sharing the same config
for both Job server and API server. To keep things simple, we change the code.

* Change sleep logic

Co-authored-by: Mattermod <mattermod@users.noreply.github.com>