MM-27487: Use PR milestone info from DB (#238)
* MM-26794: Implement a DB migration plan

This PR removes all the gorp related code which would perform DB changes
automatically from models, and implements a raw SQL file based approach to migation.

We use the golang-migrate library to implement migrations. And as the first migration,
we create the base schema of the DB itself.

More info is added in the README.

A CLI tool is also added to help with rolling back schema versions.

After this, we can slowly remove the gorp code from the codebase. Now gorp does not serve
any other purpose than just an ORM.

https://mattermost.atlassian.net/browse/MM-26794

* Fix lint errors

* Fix CircleCI config

* MM-27487: store PR milestone info in the DB

We store the milestone info in the DB and then use that to perform
the cherry-picking action. This saves a call to GitHub.

To achieve this, we need to differentiate between an issue and a PR
in the issue event handler because GitHub treats all PRs to be issues.

If a PR milestone is changed, we just update the DB with the latest PR info.

https://mattermost.atlassian.net/browse/MM-27487

* Incorporating review comments

* Remove unused comment

* Fix panic