CherryPicker Object
This commit adds a new cherrypicker object to the server package. This
object implements the proper logic to cherry-pick pull requests merged
in any fashion: squashed, merge commit, rebased.

Underneath, calls to git are still through the shell. Once we determine
this is the right approach we should consider a pure go approach,
however, go-git does not yet support cherry-picks.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@mattermost.com>