[feat] Ability to set Git merge.renamelimit variable
The number of files to consider in the exhaustive portion of rename detection during a merge. If not specified, defaults to the value of diff.renameLimit. If neither merge.renameLimit nor diff.renameLimit are specified, currently defaults to 7000. This setting has no effect if rename detection is turned off.

https://git-scm.com/docs/git-merge#Documentation/git-merge.txt-mergerenameLimit

Example of issue: https://github.com/mattermost/mattermost-server/pull/22568#issuecomment-1488555380